@@ -1,28 +1,35 @@
 {
-  "name": "my-app",
+  "name": "icss",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/icons-material": "^5.11.9",
+    "@mui/material": "^5.11.10",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.0.2",
-    "node-sass": "^6.0.0",
-    "react": "^17.0.2",
-    "react-burger-menu": "^3.0.6",
-    "react-device-detect": "^1.17.0",
-    "react-dom": "^17.0.2",
+    "moment-timezone": "^0.5.43",
+    "react": "^18.2.0",
+    "react-burger-menu": "^3.0.9",
+    "react-countdown": "^2.3.5",
+    "react-device-detect": "^2.2.3",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-lazy-load": "^3.1.13",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "^4.0.3",
-    "reactjs-countdown": "0.0.8",
-    "reactstrap": "^8.9.0",
-    "web-vitals": "^1.0.1"
+    "react-lazy-load": "^4.0.1",
+    "react-responsive": "^9.0.2",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "5.0.1",
+    "reactjs-countdown": "^0.0.8",
+    "reactstrap": "^9.2.3",
+    "sass": "^1.58.3",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "CI=false && react-scripts build",
+    "build": "set \"GENERATE_SOURCEMAP=false\" && react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
