@@ -1,18 +1,17 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import "./css/index.scss";
-import "bootstrap/dist/css/bootstrap.css";
+import ReactDOM from "react-dom/client";
+import "./index.css";
 import App from "./App";
-import reportWebVitals from "./reportWebVitals";
 
-ReactDOM.render(
-  <React.Fragment>
+if (process.env.NODE_ENV === 'development') {
+  console.log = () => {}
+  console.error = () => {}
+  console.debug = () => {}
+}
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+  <React.StrictMode>
     <App />
-  </React.Fragment>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
